First commit (major work commit)
I have tried to apply all the new technics learned, it's all about css
styling. I still need to figure out how img can change according to the
zoom in browser. And I need to arrange the css file with comments for it
to be visible and easier to debug.